amelioration de certaine chose :p

+++ src/FormationBundle/Controller/DefaultController.php
@@ -3,50 +3,45 @@
 namespace FormationBundle\Controller;
 
 use FormationBundle\Entity\Commentaire;
+use FormationBundle\Entity\Eleve;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 
 use FormationBundle\Entity\Module;
+use FormationBundle\Entity\Promotion;
+use FormationBundle\Form\PromotionType;
 use FormationBundle\Form\ModuleType;
 use FormationBundle\Entity\Note;
 use FormationBundle\Form\NoteType;
 
 class DefaultController extends Controller
 {
-    public function indexAction($id, Request $request)
+    public function indexAction( Request $request, Promotion $promotion, Eleve $ideleve)
     {
         $em = $this->getDoctrine()->getManager();
 
-
-        $promotion = $em->getRepository('FormationBundle:Promotion')->findOneBy(array('id' => $id));
-        $eleve = $em->getRepository('FormationBundle:Eleve')->findOneBy(array('id' => $id));
-        $module = $em->getRepository('FormationBundle:Module')->findAll();
-        $notes = $em->getRepository('FormationBundle:Note')->findBy(array(
-            'eleve' => $eleve,
-            'promotion' => $promotion,
-            ));
-        $commentaires = $em->getRepository('FormationBundle:Commentaire')->findById($id);
 
 
         $note = new Note();
         $form = $this->createForm('FormationBundle\Form\NoteType', $note);
         $form->handleRequest($request);
-        //var_dump($notes);die;
-        if ($form->isSubmitted())
-        {
-            foreach ($notes as $n) {
-                $del = $this->createDeleteForm($n);
-                $del->handleRequest($request);
-
-
-                $em->remove($n);
-                $em->flush();
+        if ($form->isSubmitted()) {
+            if (!empty($notes)) {
+                foreach ($notes as $n) {
+                    $formdeletenote = $this->createFormBuilder()
+                        ->setAction($this->generateUrl('note_delete', array('id' => $n->getId())))
+                        ->setMethod('DELETE')
+                        ->getForm()
+                        ->handleRequest($request);
+                    $em = $this->getDoctrine()->getManager();
+                    $em->remove($n);
+                    $em->flush();
+                }
             }
 
-            for ($i=0; $i < count($_POST['note']); $i++)
-            {
+            for ($i = 0; $i < count($_POST['note']); $i++) {
                 $note = new Note();
-                $note->setEleve($eleve);
+                $note->setEleve($ideleve);
                 $note->setPromotion($promotion);
                 $note->setNote($_POST['note'][$i]);
                 $em->persist($note);
@@ -57,18 +52,21 @@ class DefaultController extends Controller
             return $this->redirectToRoute('eleve_index', array('id' => $note->getId()));
         }
 
+        $modules = $em->getRepository('FormationBundle:Module')->findModule($ideleve);
+        $notes = $em->getRepository('FormationBundle:Note')->findBy(array('eleve' => $ideleve, 'promotion' => $promotion));
+
         return $this->render('FormationBundle:Default:index.html.twig', array(
-            'module' => $module,
-            'eleve' => $eleve,
+            'notes' => $notes,
+            'modules' => $modules,
+            'eleve' => $ideleve,
             'note' => $note,
-            'notes'=>$notes,
             'form' => $form->createView(),
             'promotion' => $promotion,
-            'commentaire' => $commentaires
         ));
     }
 
-    private function createDeleteForm(Note $note)
+
+    public function createDeleteForm(Note $note)
     {
         return $this->createFormBuilder()
             ->setAction($this->generateUrl('note_delete', array('id' => $note->getId())))
